move line client to new class
